requirements.txt
openai
python-dotenv
streamlit 
privateai_client
nemoguardrails
langchain_openai
tenacity